+++ src/core/sig.c
@@ -48,7 +48,7 @@ static int sigs_to_handle[] = { /*SIGALRM,*/ SIGHUP, SIGINT, /*SIGPIPE,*/ /*SIGP
                                 SIGSEGV, SIGTRAP, /*SIGSYS, SIGEMT,*/ SIGBUS,
                                 SIGXCPU, SIGXFSZ };
 
-#define NUM_SIGS_TO_HANDLE (sizeof(sigs_to_handle)/sizeof(sigs_to_handle[0]))
+#define NUM_SIGS_TO_HANDLE ((int)(sizeof(sigs_to_handle)/sizeof(sigs_to_handle[0])))
 
 static SigHandled sigs_handled[NUM_SIGS_TO_HANDLE];
